tests/wpt/web-platform-tests/orientation-event/t006-manual.html

<!DOCTYPE html>
<html>
  <head>
    <title>Rotate the device frame around its z axis</title>
    <meta name=viewport content="width=device-width, maximum-scale=1.0">
    <script src="/resources/testharness.js"></script>
    <script src="/resources/testharnessreport.js"></script>
    <link rel="author" title="Mosquito FP7">
    <link rel="reviewer author" title="Zhiqiang Zhang" href="mailto:zhiqiang.zhang@intel.com"> <!-- 2013-10-12 -->
  </head>
  <body>
    <p>Rotate the device frame around its z axis to run the test.</p>
    <div id="log"></div>
    <script>
      var alpha = 0;
      var run = false;
      window.addEventListener("deviceorientation", function(e) {
        if (!run) {
          run = true;
          alpha = e.alpha;
        }
      }, false);

      test(function() {
        assert_true(alpha > 0 && alpha < 360);
      }, "Alpha is between 0 and 360");

      test(function() {
        assert_false(alpha > -2 && alpha < 2);
      }, "Alpha is not around zero (between 0 and 2)");
    </script>
  </body>
</html>
